rename desktop->notmobile, none->desktop
